+++ config/initializers/redis.rb
@@ -1,1 +1,5 @@
-Redis.current = Redis.new(url: ENV.fetch("REDIS_URL") { "redis://redis:6379/1" })
+module RedisClient
+  def self.current
+    @redis ||= Redis.new(url: ENV.fetch("REDIS_URL") { "redis://redis:6379/1" })
+  end
+end